requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
filelock==3.18.0
fsspec==2025.7.0
grpcio==1.73.1
jinja2==3.1.6
joblib==1.5.1
markdown==3.8.2
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
progressbar==2.5
protobuf==6.31.1
pydantic==2.11.7
pydantic-core==2.33.2
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@2723c7fa4f2ead16ae1ce3d8afe977724bb8f87f
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
scipy==1.15.3
setuptools==80.9.0
simsimd==6.5.0
six==1.17.0
stringzilla==3.12.5
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
werkzeug==3.1.3
